(consts): Add hardware pwm channels
Previously there was no way to convert a u8 into a rppal::pwm::Channel; This meant we would've had to use rppal as a dependency in the consts crate. However with https://github.com/golemparts/rppal/issues/168 we can now define the hardware pwm channel in the consts crate, and construct the channel inside of the defaults crate, which has rppal as a dependency anyway.