(server): Improve Health Check
The health check would previously require a response from the hardware thread, which meant that excessive health checks would slow down the hardware thread. We can get rid of this overhead by simply checking if the Command channel has been closed. This tells us if incoming commands can still be processed.

The health endpoint would previously return an internal server error on the HTTP level. We likely want to change this to only on the API level later, but for now lets keep the same behaviour.